# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for Job Applier 3000 platform",
        "title": "Job Applier 3000 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/jobs/": {
            "get": {
                "description": "Retrieve all job postings",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get all jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Job"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new job posting to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Create a new job",
                "parameters": [
                    {
                        "description": "Job data",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Job"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schema.Job"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/jobs/query": {
            "get": {
                "description": "Get jobs by filtering fields like title, location, salary, company, etc.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Query jobs with filters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID (ObjectID)",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Title (regex match)",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Company ID (ObjectID)",
                        "name": "companyID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Location (regex match)",
                        "name": "location",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum salary",
                        "name": "minSalary",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum salary",
                        "name": "maxSalary",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Work type (e.g., Full-time, Part-time)",
                        "name": "workType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Work arrangement (e.g., Remote, On-site)",
                        "name": "workArrangement",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Post open date (1d or 6w)",
                        "name": "postOpenDate",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "If true, returns the latest 3 jobs",
                        "name": "latest",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting: dateAsc | dateDesc | title",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.Job"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/jobs/{id}": {
            "get": {
                "description": "Retrieve details of a specific job posting",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get a job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Job"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update a job posting by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Update a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated job data",
                        "name": "job",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.Job"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.Job"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a job posting by ID and notify applicants",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Delete a job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reason for deletion",
                        "name": "reason",
                        "in": "body",
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.Job": {
            "type": "object",
            "required": [
                "applicationDeadline",
                "companyID",
                "currency",
                "education",
                "experienceLevel",
                "jobDescription",
                "jobSummary",
                "location",
                "maxSalary",
                "minSalary",
                "numberOfPositions",
                "postOpenDate",
                "requiredSkills",
                "title",
                "visibility",
                "workArrangement",
                "workType"
            ],
            "properties": {
                "applicationDeadline": {
                    "type": "string"
                },
                "autoReject": {
                    "type": "boolean"
                },
                "companyID": {
                    "type": "string"
                },
                "currency": {
                    "type": "string",
                    "minLength": 0
                },
                "education": {
                    "type": "string",
                    "minLength": 0
                },
                "emailNotifications": {
                    "type": "boolean"
                },
                "experienceLevel": {
                    "type": "string",
                    "minLength": 0
                },
                "id": {
                    "type": "string"
                },
                "jobDescription": {
                    "type": "string",
                    "minLength": 0
                },
                "jobSummary": {
                    "type": "string",
                    "minLength": 0
                },
                "location": {
                    "type": "string",
                    "minLength": 0
                },
                "maxSalary": {
                    "type": "number",
                    "minimum": 0
                },
                "minSalary": {
                    "type": "number",
                    "minimum": 0
                },
                "niceToHave": {
                    "type": "string"
                },
                "numberOfPositions": {
                    "type": "integer"
                },
                "postOpenDate": {
                    "type": "string"
                },
                "questions": {
                    "type": "string"
                },
                "requiredSkills": {
                    "type": "string",
                    "minLength": 0
                },
                "title": {
                    "type": "string",
                    "minLength": 0
                },
                "visibility": {
                    "type": "string",
                    "minLength": 0
                },
                "workArrangement": {
                    "type": "string",
                    "minLength": 0
                },
                "workType": {
                    "type": "string",
                    "minLength": 0
                }
            }
        },
...
    }
}